[Code Revamp] Correcting necessary lib files!
