+++ train_latent_dyn_diffeq.py
@@ -10,8 +10,8 @@ from tqdm import tqdm
 from constants import Constants as const
 from lib.supervised.sgg.dsgdetr.matcher import HungarianMatcher
 from lib.supervised.sgg.dsgdetr.track import get_sequence_with_tracking
-from lib.supervised.sga import SceneSayerODE as ODE
-from lib.supervised.sga import SceneSayerSDE
+from lib.supervised.sga import scene_sayer_ode as ODE
+from lib.supervised.sga import scene_sayer_sde
 from train_base import fetch_train_basic_config, prepare_optimizer, fetch_loss_functions, save_model
